Improve `console.table()` and split out `formatTable()`
